[Fix] Cleanup compiler warnings

+++ VkRenderer/vk_loader.cpp
@@ -499,7 +499,7 @@ std::optional<std::shared_ptr<LoadedGLTF>> loadGltf(VulkanEngine *engine, std::s
             // loop the vertices of this surface, find min/max bounds
             glm::vec3 minpos = vertices[initial_vtx].position;
             glm::vec3 maxpos = vertices[initial_vtx].position;
-            for (int i = static_cast<int>(initial_vtx); i < vertices.size(); i++) {
+            for (int i = static_cast<int>(initial_vtx); i < static_cast<int>(vertices.size()); i++) {
                 minpos = glm::min(minpos, vertices[i].position);
                 maxpos = glm::max(maxpos, vertices[i].position);
             }
@@ -553,7 +553,7 @@ std::optional<std::shared_ptr<LoadedGLTF>> loadGltf(VulkanEngine *engine, std::s
     }
 
     // run loop again to setup transform hierarchy
-    for (int i = 0; i < gltf.nodes.size(); i++) {
+    for (int i = 0; i < static_cast<int>(gltf.nodes.size()); i++) {
         fastgltf::Node &node = gltf.nodes[i];
         std::shared_ptr<Node> &sceneNode = nodes[i];
